requirements.txt
streamlit
requests
transformers
torch
pdfminer.six
nltk
pandas
sentencepiece
